[CP1] Update README
Updated README with Crusine summary. In addition to explaining the project, explained how the user's information will be utilized and how to place an order within our website. Things to consider:

+ User Profile: Safety and Saving Information
+ Utilizing some information from README in About Page (Erin)

This commit is a commit made directly to the branch because it is the documentation: Pushed directly to GitHub to preserve and preview GitHub markdown text and formatting for convenience. Every thing else pushed will be via command line (Git/Git Bash) or GitHub Desktop.